Update BlockHistory elements
//BlockHistory
Added a resourcelocation to block history to determine in which dimension a block is.
//BlockHistoryHandler
Added a couple methods to simplify repeated actions. makeBlockPosFromArray and convertCompoundNBTToBlockHistory.
Changed the NBT_TAG_NAME's from private to public
Added a check to pushBlockHistory that checks that the watch hasn't gone over the set limit
//EventHandler
The push now includes the world